T2_2024_Release/Diarisation/requirements.txt
pandas
ffmpeg
SpeechRecognition
soundfile
sounddevice
openai-whisper
pyannote.audio
torchvision